# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Tickly Ticket Management API
  version: '1.0.0'
  description: API spec (MVP) for Ticket Management System — tenant-aware, RBAC enforced via JWT claims.
servers:
  - url: http://localhost:5000
    description: Local dev
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DeptDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
...
paths:
...
  /api/admin/departments:
    get:
      summary: List departments (tenant)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeptDto'
    post:
      summary: Create department
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeptDto'
...